equipment select step broken down to multiple components
