+++ src/components/common/Table.tsx
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import {
   Checkbox,
-  createStyles,
   IconButton,
   lighten,
   makeStyles,
@@ -127,7 +126,7 @@ function EnhancedTableHead<T>(props: EnhancedTableProps<T>) {
   const columns: HeadCell<T>[] = hasActions
     ? [
         {
-          id: 'id' as any,
+          id: 'id' as any, // TODO: use generic solution
           numeric: false,
           label: 'Actions',
           padding: 'default',
@@ -175,29 +174,27 @@ function EnhancedTableHead<T>(props: EnhancedTableProps<T>) {
   );
 }
 
-const useToolbarStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      paddingLeft: theme.spacing(2),
-      paddingRight: theme.spacing(1),
-    },
-    highlight:
-      theme.palette.type === 'light'
-        ? {
-            color: theme.palette.secondary.main,
-            backgroundColor: lighten(theme.palette.secondary.light, 0.9),
-          }
-        : {
-            color: theme.palette.text.primary,
-            backgroundColor: theme.palette.secondary.dark,
-          },
-    title: {
-      flex: '1 1 100%',
-      alignItems: 'center',
-      display: 'flex',
-    },
-  })
-);
+const useToolbarStyles = makeStyles((theme: Theme) => ({
+  root: {
+    paddingLeft: theme.spacing(2),
+    paddingRight: theme.spacing(1),
+  },
+  highlight:
+    theme.palette.type === 'light'
+      ? {
+          color: theme.palette.secondary.main,
+          backgroundColor: lighten(theme.palette.secondary.light, 0.9),
+        }
+      : {
+          color: theme.palette.text.primary,
+          backgroundColor: theme.palette.secondary.dark,
+        },
+  title: {
+    flex: '1 1 100%',
+    alignItems: 'center',
+    display: 'flex',
+  },
+}));
 
 export type TooltipAction = {
   tooltip: string;
